backend/app.py
from flask import Flask, jsonify, request
from flask_cors import CORS
from dotenv import load_dotenv
import os
from pymongo import MongoClient
import logging

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = Flask(__name__)
CORS(app)

# MongoDB configuration
mongo_uri = os.getenv('MONGODB_URI', 'mongodb://localhost:27017/')
client = MongoClient(mongo_uri)
db = client.turbofan_rul

@app.route('/api/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({"status": "healthy"})

@app.route('/api/predict', methods=['POST'])
def predict_rul():
    """Endpoint for RUL prediction"""
    try:
        data = request.get_json()
        # TODO: Implement prediction logic
        return jsonify({"status": "success", "prediction": None})
    except Exception as e:
        logger.error(f"Prediction error: {str(e)}")
        return jsonify({"status": "error", "message": str(e)}), 500

@app.route('/api/models', methods=['GET'])
def get_available_models():
    """Get list of available models"""
    models = [
        "LSTM",
        "Bi-LSTM",
        "Transformer",
        "CNN-LSTM",
        "LightGBM",
        "Multi-head attention LSTM"
    ]
    return jsonify({"models": models})

@app.route('/api/explain', methods=['POST'])
def explain_prediction():
    """Endpoint for model explainability"""
    try:
        data = request.get_json()
        # TODO: Implement SHAP/LIME explanation
        return jsonify({"status": "success", "explanation": None})
    except Exception as e:
        logger.error(f"Explanation error: {str(e)}")
        return jsonify({"status": "error", "message": str(e)}), 500

@app.route('/api/retrain', methods=['POST'])
def retrain_model():
    """Endpoint for model retraining"""
    try:
        data = request.get_json()
        # TODO: Implement retraining logic
        return jsonify({"status": "success", "message": "Model retraining initiated"})
    except Exception as e:
        logger.error(f"Retraining error: {str(e)}")
        return jsonify({"status": "error", "message": str(e)}), 500

if __name__ == '__main__':
    app.run(debug=True, host='0.0.0.0', port=5000) 