requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
torch==2.2.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
shap==0.42.1
lime==0.2.0.1
lightgbm==4.0.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
scipy==1.11.4 